Fix UE_LOG callsites that have format string-related UB
#rb trivial
#preflight 647107780a6634dbb236a0a6

[CL 25651638 by kirill zorin in ue5-main branch]

+++ Engine/Plugins/FX/Niagara/Source/Niagara/Private/NiagaraWorldManager.cpp
@@ -183,7 +183,7 @@ FAutoConsoleCommandWithWorldAndArgs GCmdNiagaraPlaybackRate(
 			{
 				if (Args.Num() != 1)
 				{
-					UE_LOG(LogNiagara, Log, TEXT("fx.Niagara.Debug.PlaybackRate %5.2f"), (int32)WorldManager->GetDebugPlaybackRate());
+					UE_LOG(LogNiagara, Log, TEXT("fx.Niagara.Debug.PlaybackRate %5.2f"), WorldManager->GetDebugPlaybackRate());
 				}
 				else
 				{
